BUG FIX: comparison problem between uint8_t and int on some gcc versions.

+++ userspace/src/shim6d.c
@@ -2193,7 +2193,7 @@ static int del_addr(struct in6_addr* addr, int ifidx)
 {
 	struct locset *ls, *ls_it;
 	struct hba_set* hs;
-	uint8_t valid_method=0;
+	int valid_method=0;
 	shim6_loc_l* locator;
 	int i,found=0;
 	int list_cnt;
